# fixtures/order_management/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synthetic Order Management API",
    "version": "1.0.0"
  },
  "paths": {
    "/orders": {
      "get": {
        "operationId": "list_orders",
        "summary": "List orders",
        "description": "Lists orders visible to the current user.",
        "responses": {
          "200": {
            "description": "Orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"type": "object"}
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_order",
        "summary": "Create order",
        "description": "Creates an order for a customer.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customer_id": {"type": "string"},
                  "items": {
                    "type": "array",
                    "items": {"type": "string"}
                  }
                },
                "required": ["customer_id", "items"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created order",
            "content": {
              "application/json": {
                "schema": {"type": "object"}
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}": {
      "get": {
        "operationId": "get_order",
        "summary": "Get order",
        "description": "Returns one order.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "description": "Order identifier",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Order",
            "content": {
              "application/json": {
                "schema": {"type": "object"}
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_order",
        "summary": "Delete order",
        "description": "Permanently deletes an order.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "204": {"description": "Order deleted"}
        }
      }
    },
    "/orders/{order_id}/address": {
      "patch": {
        "operationId": "update_order_address",
        "summary": "Update order address",
        "description": "Updates the delivery address for an order.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "street": {"type": "string"},
                  "city": {"type": "string"}
                },
                "required": ["street", "city"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated order",
            "content": {
              "application/json": {
                "schema": {"type": "object"}
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/refund": {
      "post": {
        "operationId": "refund_order",
        "summary": "Refund order",
        "description": "Requests a refund for an eligible order.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Refund result",
            "content": {
              "application/json": {
                "schema": {"type": "object"}
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/archive": {
      "post": {
        "operationId": "archive_order",
        "summary": "Archive order",
        "description": "Archives an order in a backend not represented by the supplied source fixture.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "202": {
            "description": "Archive accepted",
            "content": {
              "application/json": {
                "schema": {"type": "object"}
              }
            }
          }
        }
      }
    },
    "/internal/health": {
      "get": {
        "operationId": "health",
        "summary": "Internal health check",
        "description": "Reports internal service health.",
        "responses": {
          "200": {
            "description": "Service health",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {"type": "string"}
                  },
                  "required": ["status"]
                }
              }
            }
          }
        }
      }
    }
  }
}